requirements.txt
appdirs==1.4.2
colorama==0.3.7
curl==2017.0.1
Django==1.10.5
django-appconf==1.0.2
django-cleanup==0.4.2
django-cors-headers==2.0.0
django-extra-fields==0.9
django-filter==1.0.1
django-imagekit==3.3
djangorestframework==3.5.3
futures==3.0.5
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.6.0
httpie==0.9.9
jsonpickle==0.9.4
Markdown==2.6.7
mysqlclient==1.3.9
numpy==1.11.2+mkl
opencv-python==2.4.13
packaging==16.8
pigar==0.6.10
pilkit==1.1.13
Pillow==3.1.0
Pygments==2.2.0
pyparsing==2.1.10
requests==2.13.0
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
sklearn==0.0
virtualenv==15.1.0
